generator/requirements.txt
python-docx
reportlab
